#2 Temperature is below 0 degrees it's as 65530

+++ src/main/java/com/hoymiles/infrastructure/repository/mapper/Msg8717ToRealDataMapper.java
@@ -29,7 +29,7 @@ public class Msg8717ToRealDataMapper implements GenericMapper<RealDataNew.Msg871
                                 .gridReactivePower((float) src1.getQ() / 10f)
                                 .gridCurrent((float) src1.getI() / 100f)
                                 .powerFactor((float) src1.getPf() / 100f)
-                                .temp((float) src1.getTemp() / 10f)
+                                .temp((float) (short)src1.getTemp() / 10f)
                                 .link(src1.getLink())
                                 .build()
                 ).collect(Collectors.toList()))
